finish vote equivocation report, refactor code (vote -> forensics)
