+++ consensus/tests/engine_v2_tests/vote_test.go
@@ -674,52 +674,3 @@ func TestVotePoolKeepGoodHygiene(t *testing.T) {
 		}
 	}
 }
-
-func TestVoteEquivocation(t *testing.T) {
-	var numOfForks = new(int)
-	*numOfForks = 1
-	blockchain, _, currentBlock, signer, signFn, currentForkBlock := PrepareXDCTestBlockChainForV2Engine(t, 901, params.TestXDPoSMockChainConfig, &ForkedBlockOptions{numOfForkedBlocks: numOfForks})
-	engineV2 := blockchain.Engine().(*XDPoS.XDPoS).EngineV2
-	// Set round to 5
-	engineV2.SetNewRoundFaker(blockchain, types.Round(5), false)
-
-	blockInfo := &types.BlockInfo{
-		Hash:   currentBlock.Hash(),
-		Round:  types.Round(5),
-		Number: big.NewInt(901),
-	}
-	voteForSign := &types.VoteForSign{
-		ProposedBlockInfo: blockInfo,
-		GapNumber:         450,
-	}
-	voteSigningHash := types.VoteSigHash(voteForSign)
-	signedHash, err := signFn(accounts.Account{Address: signer}, voteSigningHash.Bytes())
-	assert.Nil(t, err)
-	voteMsg := &types.Vote{
-		ProposedBlockInfo: blockInfo,
-		Signature:         signedHash,
-		GapNumber:         450,
-	}
-	err = engineV2.VoteHandler(blockchain, voteMsg)
-	assert.Nil(t, err)
-	blockInfo = &types.BlockInfo{
-		Hash:   currentForkBlock.Hash(),
-		Round:  types.Round(5),
-		Number: big.NewInt(901),
-	}
-	voteForSign = &types.VoteForSign{
-		ProposedBlockInfo: blockInfo,
-		GapNumber:         450,
-	}
-	voteSigningHash = types.VoteSigHash(voteForSign)
-	signedHash, err = signFn(accounts.Account{Address: signer}, voteSigningHash.Bytes())
-	assert.Nil(t, err)
-	voteMsg = &types.Vote{
-		ProposedBlockInfo: blockInfo,
-		Signature:         signedHash,
-		GapNumber:         450,
-	}
-	err = engineV2.VoteHandler(blockchain, voteMsg)
-	assert.Nil(t, err)
-	// TODO: test forensic report
-}
